smarter url parsing
